+++ updatable_vertical_ticker_library/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.3
+- add enableSmoothScrolling option to LED variant to enable smooth pixel-precise scrolling (otherwise on false, scroll authentically on ledSize boundary) which means LED dot by dot
+
 ## 1.1.2
 \- fix addToProgress calculation (speed correction based on scrollDuration parameter)