+++ updatable_vertical_ticker_sample/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.3
+
+- add enableSmoothScrolling option to LED variant to enable smooth pixel-precise scrolling (otherwise on false, scroll authentically on ledSize boundary) which means LED dot by dot
+
 ## 1.0.1
 
 - update license to MIT